refactor(voice): restructure VoiceConfigurationPage into directory
- Move VoiceProviderSetupModal to shared.tsx with named exports
- Move VoiceDisconnectModal, getProviderIcon, getProviderLabel to shared.tsx
- Replace ConfirmationModalLayout with direct Modal (width="md")
- Use moreIcon1/moreIcon2 props instead of custom LogoArrangement
- Replace activation error Callouts with toast notifications
- Update eleven-labs logo SVG
